// src/types/film.ts
export interface Film {
  characters?: number[];
  created?: string;
  director?: string;
  edited?: string;
  episode_id?: number;
  id: number;
  opening_crawl?: string;
  planets?: number[];
  producer?: string;
  release_date?: string;
  species?: number[];
  starships?: number[];
  title: string;
  url?: string;
  vehicles?: number[];
}